More refactoring,  split responsibilities
