+++ src/main/java/nl/ictu/controller/v1/ExchangeIdentifierController.java
@@ -1,7 +1,5 @@
 package nl.ictu.controller.v1;
 
-import static nl.ictu.pseudoniemenservice.generated.server.model.WsIdentifierTypes.BSN;
-import static nl.ictu.pseudoniemenservice.generated.server.model.WsIdentifierTypes.ORGANISATION_PSEUDO;
 import static org.springframework.http.HttpStatus.UNPROCESSABLE_ENTITY;
 
 import lombok.RequiredArgsConstructor;
@@ -9,8 +7,7 @@ import lombok.SneakyThrows;
 import nl.ictu.pseudoniemenservice.generated.server.api.ExchangeIdentifierApi;
 import nl.ictu.pseudoniemenservice.generated.server.model.WsExchangeIdentifierRequest;
 import nl.ictu.pseudoniemenservice.generated.server.model.WsExchangeIdentifierResponse;
-import nl.ictu.service.v1.map.BsnPseudoMapper;
-import nl.ictu.service.v1.map.PseudoBsnMapper;
+import nl.ictu.service.v1.ExchangeIdentifierService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -19,28 +16,26 @@ import org.springframework.web.bind.annotation.RestController;
 public final class ExchangeIdentifierController implements ExchangeIdentifierApi,
         VersionOneController {
 
-    private final BsnPseudoMapper bsnPseudoMapper;
-    private final PseudoBsnMapper pseudoBsnMapper;
+    private final ExchangeIdentifierService service;
 
+    /**
+     * Exchanges an identifier based on the provided caller OIN and request data.
+     *
+     * @param callerOIN The OIN of the caller initiating the request.
+     * @param wsExchangeIdentifierForIdentifierRequest The request object containing the identifier and additional data for the exchange process.
+     * @return A ResponseEntity containing a WsExchangeIdentifierResponse if the exchange is successful,
+     *         or a ResponseEntity with HTTP status UNPROCESSABLE_ENTITY if the exchange fails.
+     */
     @Override
     @SneakyThrows
     public ResponseEntity<WsExchangeIdentifierResponse> exchangeIdentifier(final String callerOIN,
             final WsExchangeIdentifierRequest wsExchangeIdentifierForIdentifierRequest) {
 
-        final var wsIdentifierRequest = wsExchangeIdentifierForIdentifierRequest.getIdentifier();
-        final var recipientOIN = wsExchangeIdentifierForIdentifierRequest.getRecipientOIN();
-        final var recipientIdentifierType = wsExchangeIdentifierForIdentifierRequest.getRecipientIdentifierType();
-        if (BSN.equals(wsIdentifierRequest.getType()) && ORGANISATION_PSEUDO.equals(
-                recipientIdentifierType)) {
-            return ResponseEntity.ok(
-                    bsnPseudoMapper.map(wsIdentifierRequest.getValue(), recipientOIN));
-        } else if (ORGANISATION_PSEUDO.equals(wsIdentifierRequest.getType()) && BSN.equals(
-                recipientIdentifierType)) {
-            return ResponseEntity.ok(
-                    pseudoBsnMapper.map(wsIdentifierRequest.getValue(), recipientOIN));
+        final var identifier = service.exchangeIdentifier(callerOIN,
+                wsExchangeIdentifierForIdentifierRequest);
+        if (identifier != null) {
+            return ResponseEntity.ok(identifier);
         }
         return ResponseEntity.status(UNPROCESSABLE_ENTITY).build();
     }
-
-
 }
