+++ src/main/java/nl/ictu/controller/v1/GetTokenController.java
@@ -18,6 +18,14 @@ public final class GetTokenController implements GetTokenApi, VersionOneControll
 
     private final GetTokenService getTokenService;
 
+    /**
+     * Retrieves a token based on the provided caller identifier and request details.
+     *
+     * @param callerOIN The identifier of the caller organization initiating the request.
+     * @param wsGetTokenRequest The request object containing the recipient organization identifier and additional details.
+     * @return A ResponseEntity containing the token if the request is successful,
+     * or a ResponseEntity with a status of UNPROCESSABLE_ENTITY if the token cannot be retrieved.
+     */
     @Override
     @SneakyThrows
     public ResponseEntity<WsGetTokenResponse> getToken(final String callerOIN,
